chore: export all composables and unit test to validate the are exported
