+++ apps/web/composables/useCategoriesSearch/index.ts
@@ -1,0 +1,1 @@
+export * from './useCategoriesSearch';
