+++ apps/web/composables/useCategoryDetails/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './useCategoryDetails';
