+++ apps/web/composables/useCategorySettingsCollection/index.ts
@@ -1,0 +1,1 @@
+export * from './useCategorySettingsCollection';
