+++ apps/web/composables/useCategoryTemplate/homepageTemplateDataDe.json
