+++ apps/web/composables/useImageZoom/index.ts
@@ -1,0 +1,1 @@
+export * from './useImageZoom';
