+++ apps/web/composables/usePageMeta/index.ts
@@ -1,0 +1,1 @@
+export * from './usePageMeta';
