+++ apps/web/composables/usePageTitle/index.ts
@@ -1,0 +1,1 @@
+export * from './usePageTitle'
