+++ apps/web/composables/usePages/index.ts
@@ -1,0 +1,1 @@
+export * from './usePages';
