+++ apps/web/composables/useRestrictedAddress/index.ts
@@ -1,0 +1,1 @@
+export * from './useRestrictedAddress';
