+++ apps/web/composables/useUpdatePageTemplate/index.ts
@@ -1,0 +1,1 @@
+export * from './useUpdatePageTemplate';
